Use the same type in marshal and unmarshal
We marshal the length of the ciphertext as a uint32, but we were
unmarshaling it as an int32. I think this would only be a problem if the
length of the ciphertext was larger than math.MaxInt32.

In that case we'd likely get a panic when trying to call `make` with a
negative number.